added auth

+++ backend/src/controllers/notes.js
@@ -1,5 +1,15 @@
 const notesRouter = require('express').Router();
 const Note = require('../models/note');
+const User = require('../models/user');
+const jwt = require('jsonwebtoken');
+
+const getTokenFrom = request => {
+  const authorization = request.get('authorization');
+  if (authorization && authorization.toLowerCase().startsWith('bearer ')) {
+    return authorization.substring(7);
+  }
+  return null;
+};
 
 notesRouter.get('/', async (request, response) => {
   const notes = await Note.find({});
@@ -18,17 +28,41 @@ notesRouter.get('/:id', async (request, response) => {
 notesRouter.post('/', async (request, response) => {
   const body = request.body;
 
+  const token = getTokenFrom(request);
+  const decodedToken = jwt.verify(token, process.env.SECRET);
+  if (!decodedToken.id) {
+    return response.status(401).json({ error: 'token missing or invalid' });
+  }
+  const user = await User.findById(decodedToken.id);
+
   const note = new Note({
     content: body.content,
     important: body.important || false,
     date: new Date(),
+    user: user._id
   });
 
   const savedNote = await note.save();
+  user.notes = user.notes.concat(savedNote._id);
+  await user.save();
+
   response.status(201).json(savedNote);
 });
 
 notesRouter.delete('/:id', async (request, response) => {
+  const token = getTokenFrom(request);
+  const decodedToken = jwt.verify(token, process.env.SECRET);
+  if (!decodedToken.id) {
+    return response.status(401).json({ error: 'token missing or invalid' });
+  }
+  const user = await User.findById(decodedToken.id);
+  const note = await Note.findById(request.params.id);
+  console.log(`User (${user.id}) Note(${note.user}) isEqual(${user.id.toString() === note.user.toString()})`);
+  if (user.id.toString() !== note.user.toString()) {
+    response.status(403).json({ error: 'no permissions to delete this note' });
+    return;
+  }
+
   await Note.findByIdAndRemove(request.params.id);
   response.status(204).end();
 });
